requirements.txt
python-docx>=1.2.0
reportlab>=4.0.0
pydantic>=2.0.0
click>=8.1.0
openai>=1.0.0
flask>=3.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
ruff>=0.1.0
mypy>=1.0.0